vm: allow accessing non-existent keys in objects using `[]`
Previously, `{}["foo"]` would return an error for the key "foo" not
existing in the object.

This commit loosens that restriction, and now returns `null` for any key
that does not exist in an object when the `[]` operator is used to
access the key.

The `.` operator will still return an error for non-existent keys, such
as `{}.foo`.

Related to grafana/agent#5615.